Move JsonConverters to better place
Move the new json converters to the models' folder where they are
closer to the serilizable models and prevent mixing the new converters
with the old converters.

Additionally move the list of converters to the
`JsonSerializationOptions` class in order to share the exact same
settings in the client and server code.
