+++ WalletWasabi/WabiSabi/Models/Serialization/JsonSerializationOptions.cs
@@ -1,0 +1,32 @@
+using System;
+using System.Collections.Generic;
+using Newtonsoft.Json;
+using WalletWasabi.JsonConverters;
+using WalletWasabi.JsonConverters.Bitcoin;
+using WalletWasabi.WabiSabi.Crypto.Serialization;
+
+namespace WalletWasabi.WabiSabi.Models.Serialization
+{
+	public class JsonSerializationOptions
+	{
+		private static readonly JsonSerializerSettings settings = new ()
+		{
+			Converters = new List<JsonConverter>()
+			{
+				new ScalarJsonConverter(),
+				new GroupElementJsonConverter(),
+				new OutPointJsonConverter(),
+				new WitScriptJsonConverter(),
+				new ScriptJsonConverter(),
+				new OwnershipProofJsonConverter(),
+			}
+		};
+		public static readonly JsonSerializationOptions Default = new();
+
+		private JsonSerializationOptions()
+		{
+		}
+
+		public JsonSerializerSettings Settings => settings;
+	}
+}
