+++ WalletWasabi/WabiSabi/Models/Serialization/WitScriptJsonConverter.cs
@@ -2,24 +2,21 @@ using NBitcoin;
 using Newtonsoft.Json;
 using System;
 
-namespace WalletWasabi.JsonConverters
+namespace WalletWasabi.WabiSabi.Models.Serialization
 {
 	public class WitScriptJsonConverter : JsonConverter
 	{
-		/// <inheritdoc />
 		public override bool CanConvert(Type objectType)
 		{
 			return objectType == typeof(WitScript);
 		}
 
-		/// <inheritdoc />
 		public override object ReadJson(JsonReader reader, Type objectType, object existingValue, JsonSerializer serializer)
 		{
 			var value = (string)reader.Value;
 			return new WitScript(ByteHelpers.FromHex(value));
 		}
 
-		/// <inheritdoc />
 		public override void WriteJson(JsonWriter writer, object value, JsonSerializer serializer)
 		{
 			var bytes = ((WitScript)value).ToBytes();
